CIAS30-848 Invite researcher to the team (#147)
* CIAS30-848 Invite researcher to the team

Add possibility to add confirmed researcher to the team. We will not let
to invite not confirmed researcher account.
Super admin can add any confirmed researcher to any team.
Team admin can add any confirmed researcher only to his team.

Researchers can be added to the team by entering their email address.
When researcher does not exists in the system, then new researcher
account is created and confirmation email is sent to that email.

In case when researcher already exists in the system researcher will
receive team invitation email with invitation_token in the url. After
successful invitation acceptation he will be redirected to the Web App
with success message.
In case when invitation is already accepted, or the token is wrong, it
will redirect user to the web app with error message.

* Add Base64 encoding for messages

* Add dependent destroy to team invitations