CIAS30-910 Generate Reports based on answers (#181)
* CIAS30-910 Generate Reports based on answers

Generate reports base on user answers when the user session is finished.

The report is always generated for the user session's. It will be
available for the researcher later.

For the guest/participant it's possible to get his report based on his
answer in Participant Screen, to implement later. But if participant
will decide to not receive report it will not be shown in his dashboard.

Each report will be generated from the variants that fulfill the
formula provided in the report section. Only first variant that match
the formula will be displayed.

Name variable will be used in the variants content

* Fix test setup

* Code review fixes